requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.0
transformers==4.33.2
torch==2.0.1
python-dotenv==1.0.0
requests==2.31.0
python-magic==0.4.27
matplotlib==3.7.2
seaborn==0.12.2